Added non responsive sensor emergency stop
Added an 'emergency stop' feature to stop a run if the temperature
sensor is no longer providing data.  If the temperature sensor values
are remaining stable and exactly the same for 20 time cycles
consecutively while the heat is on, the run is reset to prevent damage
to the part or the oven.